refactor: InputConverter 유틸 클래스 구현
- 문자열을 LocalTime으로 변환하는 메서드를 분리
