+++ src/main/java/AttendanceCheck.java
@@ -1,7 +1,6 @@
 import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 
 public class AttendanceCheck {
 
@@ -42,16 +41,6 @@ public class AttendanceCheck {
 
         return "일요일";
 
-    }
-
-    public static LocalTime convertToLocalTime(String timeInput) {
-        try {
-            DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("HH:mm");
-
-            return LocalTime.parse(timeInput, dateTimeFormatter);
-        } catch (Exception e) {
-            throw new IllegalArgumentException("[ERROR] 입력된 시간 형식이 적절하지 않습니다.");
-        }
     }
 
     public static LocalTime getStandardTime(LocalDate date) {
